validationtesting: refactor testutils
