+++ testutil/testutil.go
@@ -21,9 +21,6 @@ import (
 	"math/rand"
 	"net/http"
 	"net/http/httptest"
-	"os/exec"
-	"path/filepath"
-	"strings"
 )
 
 // ContainsString checks if an array contains a string.
@@ -76,13 +73,4 @@ func RandomKey() []byte {
 	b := make([]byte, c)
 	rand.Read(b)
 	return b
-}
-
-// CompileGoPlugin calls the go compiler to build a plugin binary from a go source file.
-func CompileGoPlugin(pluginsPath, pluginName string) (string, error) {
-	sourceFile := filepath.Join(pluginsPath, pluginName)
-	pluginFile := filepath.Join(pluginsPath, strings.Replace(pluginName, ".go", ".so", 1))
-
-	cmd := exec.Command("go", "build", "-o", pluginFile, "-buildmode=plugin", sourceFile)
-	return pluginFile, cmd.Run()
 }
