--- validation/testutils/plugins/byzantine_validator.go
@@ -1,27 +1,0 @@
-// Copyright 2017 Stratumn SAS. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"errors"
-
-	"github.com/stratumn/go-chainscript"
-	"github.com/stratumn/go-core/store"
-)
-
-// Init validates the transition towards the "init" state
-func Init(storeReader store.SegmentReader, l *chainscript.Link) error {
-	return errors.New("error")
-}
