+++ validation/validationtesting/config.go
@@ -1,0 +1,30 @@
+// Copyright 2017 Stratumn SAS. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package validationtesting
+
+import "fmt"
+
+// These are the validation configuration exported by this package.
+var (
+	ValidAuctionJSONConfig = CreateValidatorJSON("auction", ValidAuctionJSONPKIConfig, ValidAuctionJSONTypesConfig)
+	ValidChatJSONConfig    = CreateValidatorJSON("chat", ValidChatJSONPKIConfig, ValidChatJSONTypesConfig)
+	ValidJSONConfig        = fmt.Sprintf(`{%s,%s}`, ValidAuctionJSONConfig, ValidChatJSONConfig)
+)
+
+// CreateValidatorJSON formats a PKI and a types schema into a valid JSON
+// configuration.
+func CreateValidatorJSON(name, pki, types string) string {
+	return fmt.Sprintf(`"%s": {"pki": %s,"types": %s}`, name, pki, types)
+}
